In GLRenderer, actually read settings from the processor for fixed size buffers
Now, if the fixed size buffer button is un-ticked for a shader this will render straight to the back buffer. Otherwise, it will render to a single off-screen texture big enough to satisfy all the shaders' needs. If the size for a shader is less than the biggest in the list, this will render to a small portion of the off-screen texture.