Deal with the unpredictability of processBlock
processBlock is not called at a regular interval. Because of this midi messages can't arrive in sync with rendering - for example, a note on event and a note off event may arrive at the same time, nullifying them. To deal with this, keep a "delay buffer" of midi messages, where we provide iKeyDown / iKeyUp based on the state of the keyboard some time in the past. This leads to far more predictable behavior.